update edit wallets entry
