+++ packages/app-extension/src/components/Unlocked/Settings/YourAccount/EditWallets/index.tsx
@@ -9,127 +9,8 @@ import { Typography } from "@mui/material";
 
 import { WalletAddress } from "../../../../common";
 import { useNavStack } from "../../../../common/Layout/NavStack";
-import { ImportTypeBadge } from "../../../../common/WalletList";
+import { ImportTypeBadge, WalletList } from "../../../../common/WalletList";
 import { AddConnectWalletButton } from "../..";
-
-export function EditWallets({
-  filter,
-}: {
-  filter?: (w: {
-    name: string;
-    publicKey: string;
-    blockchain: string;
-  }) => boolean;
-}) {
-  const nav = useNavStack();
-  const blockchainKeyrings = useWalletPublicKeys();
-
-  useEffect(() => {
-    const title = nav.title;
-    nav.setTitle("Edit wallets");
-
-    return () => {
-      nav.setTitle(title);
-    };
-  }, []);
-
-  return (
-    <div style={{ paddingTop: "16px", height: "100%" }}>
-      {Object.entries(blockchainKeyrings).map(([blockchain, keyring]) => (
-        <WalletList
-          key={blockchain}
-          blockchain={blockchain as Blockchain}
-          keyring={keyring}
-          filter={filter}
-        />
-      ))}
-      {/* Hack to add margin bottom. */}
-      <div style={{ height: "16px" }} />
-    </div>
-  );
-}
-
-function WalletList({
-  blockchain,
-  keyring,
-  filter,
-}: {
-  blockchain: Blockchain;
-  keyring: any;
-  filter?: (w: {
-    name: string;
-    publicKey: string;
-    blockchain: string;
-  }) => boolean;
-}) {
-  const theme = useCustomTheme();
-  let flattenedWallets = [
-    ...keyring.hdPublicKeys.map((k: any) => ({
-      ...k,
-      type: "derived",
-      blockchain,
-    })),
-    ...keyring.importedPublicKeys.map((k: any) => ({
-      ...k,
-      blockchain,
-      type: "imported",
-    })),
-    ...keyring.ledgerPublicKeys.map((k: any) => ({
-      ...k,
-      type: "ledger",
-      blockchain,
-    })),
-  ];
-  if (filter) {
-    flattenedWallets = flattenedWallets.filter(filter);
-  }
-
-  // TODO: replace placeholder wallet avatar with stored image when available
-  return flattenedWallets.length === 0 ? (
-    <></>
-  ) : (
-    <div style={{ marginBottom: "16px" }}>
-      <Typography
-        style={{
-          marginLeft: "16px",
-          marginRight: "16px",
-          marginBottom: "12px",
-          color: theme.custom.colors.fontColor,
-        }}
-      >
-        {toTitleCase(blockchain)}
-      </Typography>
-
-      <List
-        style={{
-          border: `${theme.custom.colors.borderFull}`,
-          borderRadius: "10px",
-        }}
-      >
-        {flattenedWallets.map(({ name, publicKey, type }, idx) => (
-          <WalletListItem
-            blockchain={blockchain}
-            key={publicKey.toString()}
-            name={name}
-            publicKey={publicKey}
-            type={type}
-            isFirst={idx === 0}
-            isLast={idx === flattenedWallets.length - 1}
-            showDetailMenu={true}
-          />
-        ))}
-      </List>
-      <div
-        style={{
-          marginLeft: "16px",
-          marginRight: "16px",
-        }}
-      >
-        <AddConnectWalletButton blockchain={blockchain} />
-      </div>
-    </div>
-  );
-}
 
 export const WalletListItem: React.FC<{
   blockchain: Blockchain;
